requirements.txt
flask
gunicorn
requests
flask-cors
mysql-connector-python
pandas
chardet
BeautifulSoup4
html5lib
lxml
xmltodict
jwt
scp
paramiko
flask_limiter
warcat
jwt
PyJWT
werkzeug